requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
pkg_resources==0.0.0
python-dotenv==0.21.1
pytz==2022.7.1
sqlparse==0.4.3
